# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Portfolio Platform API",
        "description": "Backend API for the AI-enabled dynamic portfolio platform",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:9100",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/admin/settings/profile": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Profile photo and resume",
                "description": "Admin view of the same singleton row.",
                "operationId": "adminProfile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSiteProfileResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Set profile photo and resume",
                "description": "References media uploaded earlier; null clears without deleting the file.",
                "operationId": "updateProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SiteProfileUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSiteProfileResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/articles/{id}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get an article",
                "description": "Any status.",
                "operationId": "get_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseArticleResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Blog"
                ],
                "summary": "Replace an article",
                "description": "The tags sent become the complete set.",
                "operationId": "update_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseArticleResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Blog"
                ],
                "summary": "Delete an article",
                "description": "Soft delete.",
                "operationId": "delete_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/admin/articles": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "List articles",
                "description": "All statuses, paginated, optional ?status= filter.",
                "operationId": "list_6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "SCHEDULED",
                                "PUBLISHED",
                                "ARCHIVED"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseArticleSummaryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Create an article",
                "description": "Content is sanitized on write.",
                "operationId": "create_5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ArticleCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List published projects",
                "description": "Ordered by display order. Not paginated. Optional ?search= matches title, description and technology names (FR-18).",
                "operationId": "listPublished_1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListProjectSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/media/{id}/content": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Serve a file",
                "description": "Public: the Public Site renders these to anonymous visitors.",
                "operationId": "content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/articles": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "List published articles",
                "description": "Newest first. A scheduled article appears on its own when its time passes.",
                "operationId": "listPublic",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponseArticleSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/articles/{slug}": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Get a published article",
                "description": "404 when unknown, draft, or not yet due.",
                "operationId": "getPublic",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseArticleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SiteProfileUpdateRequest": {
                "type": "object",
                "properties": {
                    "profileImageMediaId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Media id, or null to clear"
                    },
                    "resumeMediaId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Media id, or null to clear"
                    }
                }
            },
            "ApiResponseSiteProfileResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/SiteProfileResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "SiteProfileResponse": {
                "type": "object",
                "properties": {
                    "profileImage": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "resume": {
                        "$ref": "#/components/schemas/MediaResponse"
                    }
                }
            },
...
            "ArticleUpdateRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 0
                    },
                    "slug": {
                        "type": "string",
                        "description": "Optional. Derived from the title when omitted.",
                        "maxLength": 270,
                        "minLength": 0
                    },
                    "excerpt": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "content": {
                        "type": "string",
                        "minLength": 1
                    },
                    "thumbnailMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "ogImageMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category": {
                        "type": "string",
                        "description": "Category name; created if unknown",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "tags": {
                        "type": "array",
                        "description": "Tag names; unknown ones are created",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "description": "Defaults to DRAFT",
                        "enum": [
                            "DRAFT",
                            "SCHEDULED",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Required for SCHEDULED/PUBLISHED; a future value schedules it"
                    },
                    "seoTitle": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "seoDescription": {
                        "type": "string",
                        "maxLength": 300,
                        "minLength": 0
                    },
                    "aiVisible": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "content",
                    "title"
                ]
            },
            "ApiResponseArticleResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ArticleResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ArticleResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "excerpt": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "ogImage": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "category": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "SCHEDULED",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "readingTimeMinutes": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "seoTitle": {
                        "type": "string"
                    },
                    "seoDescription": {
                        "type": "string"
                    },
                    "aiVisible": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ArticleCreateRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 250,
                        "minLength": 0
                    },
                    "slug": {
                        "type": "string",
                        "description": "Optional. Derived from the title when omitted.",
                        "maxLength": 270,
                        "minLength": 0
                    },
                    "excerpt": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "content": {
                        "type": "string",
                        "minLength": 1
                    },
                    "thumbnailMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "ogImageMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "category": {
                        "type": "string",
                        "description": "Category name; created if unknown",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "tags": {
                        "type": "array",
                        "description": "Tag names; unknown ones are created",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "description": "Defaults to DRAFT",
                        "enum": [
                            "DRAFT",
                            "SCHEDULED",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Required for SCHEDULED/PUBLISHED; a future value schedules it"
                    },
                    "seoTitle": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "seoDescription": {
                        "type": "string",
                        "maxLength": 300,
                        "minLength": 0
                    },
                    "aiVisible": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "content",
                    "title"
                ]
            },
...
            "ApiResponseListProjectSummaryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectSummaryResponse"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponsePageResponseArticleSummaryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponseArticleSummaryResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}